requirements.txt
chainer
chainercv
flask
numpy
pillow
